+++ mpc_energy/CHANGELOG.md
@@ -2,6 +2,7 @@
 * Improved EV charging logic. 
 * Stopped effective price from going to zero overnight.
 * Changed SigEnergy Charger to not require additional helper.
+* Improved Effective Price and Next Grid Interaction calculations.
 
 ## 0.8.0
 * **!!!Please enter the general config page to sync old config through HA with new web dashboard config!!!**.